requirements.txt
Pillow==12.1.1
pyserial==3.5
